+++ libs/libc/mutex/Make.defs
@@ -1,0 +1,31 @@
+############################################################################
+# libs/libc/mutex/Make.defs
+#
+# Licensed to the Apache Software Foundation (ASF) under one or more
+# contributor license agreements.  See the NOTICE file distributed with
+# this work for additional information regarding copyright ownership.  The
+# ASF licenses this file to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance with the
+# License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
+# License for the specific language governing permissions and limitations
+# under the License.
+#
+############################################################################
+
+# Add the mutex C files to the build
+
+CSRCS += mutex_breaklock.c mutex_destroy.c mutex_init.c
+CSRCS += mutex_is_locked.c mutex_is_hold.c mutex_lock.c
+CSRCS += mutex_reset.c mutex_restorelock.c mutex_timedlock.c
+CSRCS += mutex_trylock.c mutex_unlock.c
+
+# Add the mutex directory to the build
+
+DEPPATH += --dep-path mutex
+VPATH += :mutex
