+++ libs/libc/mutex/mutex_restorelock.c
@@ -1,0 +1,85 @@
+/****************************************************************************
+ * libs/libc/mutex/mutex_restorelock.c
+ *
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.  The
+ * ASF licenses this file to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance with the
+ * License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ *
+ ****************************************************************************/
+
+/****************************************************************************
+ * Included Files
+ ****************************************************************************/
+
+#include <nuttx/mutex.h>
+
+/****************************************************************************
+ * Public Functions
+ ****************************************************************************/
+
+/****************************************************************************
+ * Name: nxmutex_restorelock
+ *
+ * Description:
+ *   This function attempts to restore the mutex.
+ *
+ * Parameters:
+ *   mutex   - mutex descriptor.
+ *   locked  - true: it's mean that the mutex is broke success
+ *
+ * Return Value:
+ *   This is an internal OS interface and should not be used by applications.
+ *   It follows the NuttX internal error return policy:  Zero (OK) is
+ *   returned on success.  A negated errno value is returned on failure
+ *
+ ****************************************************************************/
+
+int nxmutex_restorelock(FAR mutex_t *mutex, bool locked)
+{
+  return locked ? nxmutex_lock(mutex) : OK;
+}
+
+/****************************************************************************
+ * Name: nxrmutex_restorelock
+ *
+ * Description:
+ *   This function attempts to restore the recursive mutex.
+ *
+ * Parameters:
+ *   rmutex - Recursive mutex descriptor.
+ *   count  - Count after restore.
+ *
+ * Return Value:
+ *   This is an internal OS interface and should not be used by applications.
+ *   It follows the NuttX internal error return policy:  Zero (OK) is
+ *   returned on success.  A negated errno value is returned on failure.
+ *   Possible returned errors:
+ *
+ ****************************************************************************/
+
+int nxrmutex_restorelock(FAR rmutex_t *rmutex, unsigned int count)
+{
+  int ret = OK;
+
+  if (count != 0)
+    {
+      ret = nxmutex_lock(&rmutex->mutex);
+      if (ret >= 0)
+        {
+          rmutex->count = count;
+        }
+    }
+
+  return ret;
+}
