+++ syscall/syscall.csv
@@ -74,6 +74,12 @@
 "nx_pthread_create","nuttx/pthread.h","!defined(CONFIG_DISABLE_PTHREAD)","int","pthread_trampoline_t","FAR pthread_t *","FAR const pthread_attr_t *","pthread_startroutine_t","pthread_addr_t"
 "nx_pthread_exit","nuttx/pthread.h","!defined(CONFIG_DISABLE_PTHREAD)","noreturn","pthread_addr_t"
 "nx_vsyslog","nuttx/syslog/syslog.h","","int","int","FAR const IPTR char *","FAR va_list *"
+"nxmutex_breaklock","nuttx/mutex.h","","int","FAR mutex_t *","FAR bool *"
+"nxmutex_destroy","nuttx/mutex.h","","int","FAR mutex_t *"
+"nxmutex_lock","nuttx/mutex.h","","int","FAR mutex_t *"
+"nxmutex_timedlock","nuttx/mutex.h","","int","FAR mutex_t *","unsigned int"
+"nxmutex_trylock","nuttx/mutex.h","","int","FAR mutex_t *"
+"nxmutex_unlock","nuttx/mutex.h","","int","FAR mutex_t *"
 "nxsched_get_stackinfo","nuttx/sched.h","","int","pid_t","FAR struct stackinfo_s *"
 "open","fcntl.h","","int","FAR const char *","int","...","mode_t"
 "pgalloc", "nuttx/arch.h", "defined(CONFIG_BUILD_KERNEL)", "uintptr_t", "uintptr_t", "unsigned int"
